refactor: centralize ParsedParams interface
